pyproject.toml
[tool.poetry]
name = "ob3-document-processing-service"
version = "0.1.0"
description = "Django REST API сервис для автоматической обработки, классификации и валидации загружаемых документов"
authors = ["Rinat Gabitov riga1897@yandex.ru"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.29.0"
pillow = "^12.0.0"
pypdf = "^6.4.0"
python-docx = "^1.1.0"
django-filter = "^25.2"
django-redis = "^6.0.0"
python-decouple-stubs = "^3.8.0.4"
django-filter-stubs = "^0.1.3"
dj-database-url = "^3.0.1"
structlog = "^25.5.0"
reportlab = "^4.4.5"
gunicorn = "^23.0.0"
# filetype: кросс-платформенная библиотека для определения MIME-типов
# Чистый Python, без системных зависимостей, работает на всех платформах
filetype = "^1.2.0"
poetry-core = "^2.2.1"
celery = "^5.5.3"
redis = "^7.1.0"

[tool.poetry.group.lint.dependencies]
ruff = "^0.1.14"
mypy = "^1.8.0"
black = "^24.1.1"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
types-redis = "^4.6.0"
pre-commit = "^3.6.0"
pyright = "^1.1.407"
celery-types = "^0.23.0"

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit (security)
    "T20", # flake8-print
]
ignore = [
    "E501",  # line too long (handled by black)
]
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
    "venv",
    "*.egg-info",
    "Docker",
    "attached_assets",
    ".pythonlibs",
    "scripts",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101", "S108"]  # allow asserts and /tmp paths in tests
"tests/factories.py" = ["N805"]  # Factory Boy naming convention
"config/settings/*.py" = ["F403", "F405"]  # Django settings star imports
"config/settings/test.py" = ["S105"]  # test settings hardcoded secrets OK

[tool.ruff.isort]
known-first-party = ["apps", "config"]
section-order = ["future", "standard-library", "django", "third-party", "first-party", "local-folder"]

[tool.ruff.isort.sections]
django = ["django"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | migrations
  | __pycache__
)/
'''

[tool.mypy]
strict = true
python_version = "3.12"
mypy_path = ["."]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
disallow_untyped_decorators = false
warn_unused_ignores = false
explicit_package_bases = true
namespace_packages = true

# Django migrations - auto-generated code, ignore all errors
[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

# Third-party test fixtures - allow untyped defs but catch syntax errors
[[tool.mypy.overrides]]
module = ["factory.*", "faker.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false

# Libraries without type stubs
[[tool.mypy.overrides]]
module = ["filetype", "filetype.*", "reportlab.*"]
ignore_missing_imports = true

# Tests - Relax Factory Boy errors (untyped factories + attr-defined)
# Factory Boy factories return objects but mypy sees factory classes

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = true
disallow_incomplete_defs = false
disallow_untyped_calls = false
disable_error_code = ["attr-defined"]

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
pytest-env = "^1.2.0"
pytest-xdist = "^3.5.0"
factory-boy = "^3.3.0"
faker = "^22.6.0"

[tool.coverage.run]
data_file = "var/coverage/.coverage"
source = ["apps", "config"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "apps/core/cache.py",
]

[tool.coverage.html]
directory = "var/coverage/htmlcov"

[tool.coverage.xml]
output = "var/coverage/coverage.xml"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

# pytest-env: Force override DJANGO_SETTINGS_MODULE (ignores Replit env var)
[tool.pytest_env]
DJANGO_SETTINGS_MODULE = "config.settings.test"
PYTEST_RUNNING = "1"
